Adding a Code of Conduct
Note that I am using my company's work email as the contact, here. Another email should also be set up so that there is another email to ensure that there are multiple ways to report infractions. community@ or opensource@kontain should work fine, and can normally be set up as an alias to the relevant people. Let's make sure to have that before mergning this.