Remove Getting Started Doc (#7)
At this stage, I don't know if there is a benefit to having Getting Started in  another doc. I've moved most of the content here, except for two sections - Dev-guide and Symlinks, both of which were filled with todos and which I coudn't really ascertain the worth of.

I've added a ToC as well. Let me know what you think.

Next step is to work on the wording and to go through this guide with a fine comb.